Integración de Backend eventos con Fronted eventos

+++ 10-calendar/src/hooks/useCalendarStore.js
@@ -1,5 +1,8 @@
 import { useDispatch, useSelector } from "react-redux";
-import { onAddNewEvent, onDeleteEvent, onSetActiveEvent, onUpdateEvent } from "../store";
+import { onAddNewEvent, onDeleteEvent, onLoadEvents, onSetActiveEvent, onUpdateEvent } from "../store";
+import { calendarApi } from "../api";
+import { convertEventsToDateEvents } from "./convertEventsToDateEvents";
+import Swal from "sweetalert2";
 
 
 export const useCalendarStore = () => {
@@ -7,6 +10,7 @@ export const useCalendarStore = () => {
     const dispatch = useDispatch();
 
     const { events, activeEvent } = useSelector(state => state.calendar);
+    const { user } = useSelector(state => state.auth);
 
     const setActiveEvent = (calendarEvent) => {
         dispatch(onSetActiveEvent(calendarEvent))
@@ -14,18 +18,57 @@ export const useCalendarStore = () => {
 
     const startSavingEvent = async (calendarEvent) => {
 
+        try {
+            
+            if (calendarEvent.id) {
+                // Actualizando
+                await calendarApi.put( `/events/${calendarEvent.id}`, calendarEvent );
+                dispatch(onUpdateEvent({ ...calendarEvent, user }))
+                return;
+            }
+            // Creando
+            const { data } = await calendarApi.post('/events', calendarEvent);
+            dispatch(onAddNewEvent({ ...calendarEvent, id: data.evento.id, user }));
+            
+        } catch (error) {
+            console.log(error);
+            Swal.fire('Error al Guardar', error.response.data?.msg, 'error')
+        }
+        
 
-        if (calendarEvent._id) {
-            dispatch(onUpdateEvent({ ...calendarEvent }))
-        } else {
-            dispatch(onAddNewEvent({ ...calendarEvent, _id: new Date().getTime() }));
+    }
 
+    const startDeletingEvent = async() => {
+
+        try {
+         
+            if (!activeEvent.id) return;
+                await calendarApi.delete(`/events/${activeEvent.id}`);
+                dispatch(onDeleteEvent())  
+                
+        } catch (error) {
+            console.log(error);
+            Swal.fire('Error al Borrar', error.response.data?.msg, 'error')
+        }
+        
+    }
+
+    const startLoadingEvents = async() => {
+        try {
+            
+            const { data } = await calendarApi.get('/events')
+            const events = convertEventsToDateEvents( data.eventos )
+            dispatch(onLoadEvents(events))
+            console.log(events);
+
+
+        } catch (error) {
+            console.log('Error cargando eventos');
+            console.log(error);
+            
         }
     }
 
-    const startDeletingEvent = ( ) => {
-        dispatch(onDeleteEvent())
-    }
 
     return {
         // * Propiedades
@@ -34,9 +77,10 @@ export const useCalendarStore = () => {
         hasEventSelected: !!activeEvent,
 
         // * Métodos
-        startDeletingEvent,
         setActiveEvent,
-        startSavingEvent,
+        startDeletingEvent,
+        startLoadingEvents,
+        startSavingEvent
     }
 
 }
